fix: {<font|color|text>} not working
