+++ sushi/sushi-compose/src/commonMain/kotlin/com/zomato/sushi/compose/markdown/MarkdownParser.kt
@@ -48,8 +48,8 @@ class MarkdownParser private constructor(
                 .processor(BoldProcessor())
                 .processor(ItalicProcessor())
                 .processor(StrikethroughProcessor())
-                .processor(ColorProcessor())
                 .processor(FontWeightProcessor())
+                .processor(TextColorProcessor())
                 .processor(LinkProcessor())
                 .processor(UnderlineAnnotaterProcessor())
                 .build()
